+++ mocapi-server/src/test/java/com/callibrity/mocapi/server/autoconfigure/MocapiServerToolsAutoConfigurationTest.java
@@ -71,7 +71,7 @@ class MocapiServerToolsAutoConfigurationTest {
 
     @Bean
     MethodInvokerFactory methodInvokerFactory() {
-      return new DefaultMethodInvokerFactory(java.util.List.of());
+      return new DefaultMethodInvokerFactory();
     }
   }