+++ mocapi-server/src/test/java/com/callibrity/mocapi/server/tools/CallToolHandlersTest.java
@@ -77,7 +77,7 @@ class CallToolHandlersTest {
 
     @Bean
     MethodInvokerFactory methodInvokerFactory() {
-      return new DefaultMethodInvokerFactory(List.of());
+      return new DefaultMethodInvokerFactory();
     }
   }